move prettier/util.ts xrun function to util/xrun.ts
