+++ format.ts
@@ -2,7 +2,7 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 const { exit, args, execPath } = Deno;
 import { parse } from "./flags/mod.ts";
-import { xrun } from "./prettier/util.ts";
+import { xrun } from "./util/xrun.ts";
 
 async function main(opts): Promise<void> {
   const args = [
